+++ src/main/java/com/simisinc/platform/presentation/controller/WidgetOutputStream.java
@@ -16,16 +16,18 @@
 
 package com.simisinc.platform.presentation.controller;
 
-import org.apache.commons.io.output.WriterOutputStream;
-
-import javax.servlet.ServletOutputStream;
-import javax.servlet.WriteListener;
 import java.io.BufferedOutputStream;
 import java.io.CharArrayWriter;
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+
+import javax.servlet.ServletOutputStream;
+import javax.servlet.WriteListener;
+
+import org.apache.commons.io.output.WriterOutputStream;
 
 /**
- * Description
+ * Widget content from the server
  *
  * @author matt rajkowski
  * @created 4/6/18 10:51 AM
@@ -34,8 +36,12 @@ public class WidgetOutputStream extends ServletOutputStream {
 
   private final BufferedOutputStream bufferedOut;
 
-  public WidgetOutputStream(CharArrayWriter charArray) {
-    this.bufferedOut = new BufferedOutputStream(new WriterOutputStream(charArray, "UTF-8"), 16384);
+  public WidgetOutputStream(CharArrayWriter charArray) throws IOException {
+    this.bufferedOut = new BufferedOutputStream(WriterOutputStream.builder()
+        .setWriter(charArray)
+        .setBufferSize(16384)
+        .setCharset(StandardCharsets.UTF_8)
+        .get());
   }
 
   @Override