+++ src/main/java/com/simisinc/platform/presentation/controller/XMLFooterLoader.java
@@ -69,6 +69,7 @@ public class XMLFooterLoader implements Serializable {
       throws FactoryConfigurationError, ParserConfigurationException, SAXException, IOException {
     DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
     factory.setFeature("http://apache.org/xml/features/disallow-doctype-decl", true);
+    factory.setFeature(XMLConstants.FEATURE_SECURE_PROCESSING, true);
     factory.setXIncludeAware(false);
     factory.setExpandEntityReferences(false);
 
@@ -82,6 +83,7 @@ public class XMLFooterLoader implements Serializable {
       throws FactoryConfigurationError, ParserConfigurationException, SAXException, IOException {
     DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
     factory.setFeature("http://apache.org/xml/features/disallow-doctype-decl", true);
+    factory.setFeature(XMLConstants.FEATURE_SECURE_PROCESSING, true);
     factory.setXIncludeAware(false);
     factory.setExpandEntityReferences(false);